change REPL quit to stop
Since quit is a common command to exit certain
types of command shells, I felt it would be better
to use a different name for stopping the cluster,
since it would be easy to get confused and accidentally
shut down the cluster when simply wanting to exit the
REPL shell.

I switched the REPL command to stop() for a graceful
shutdown and a kill() for stopping hard.