Be more resilient, but only up to a point
If the connection to iriscouch hiccups, it can tank the npm-www
website.  A worker will die, then restart badly, and the master
would just give up too quickly and end up in a state where it
had no workers, and was not serving any requests.

Instead, be a little more pushy about trying to restart, and then
kill the *master* if it fails repeatedly, so that the crash is
reported to the system, and SMF/upstart/whatever can take over.